Move pattern, complex, Literal and union tests to tests/types/
Union tests are appended to the existing tests/types/test_union.py.
The `constr(pattern=...)` usage is replaced with the explicit
annotated form using `StringConstraints`.
