+++ tests/test_types.py
@@ -775,157 +775,6 @@ def test_json_required():
         JsonRequired()
 
 
-@pytest.mark.parametrize(
-    ('pattern_type', 'pattern_value', 'matching_value', 'non_matching_value'),
-    [
-        pytest.param(re.Pattern, r'^whatev.r\d$', 'whatever1', ' whatever1', id='re.Pattern'),
-        pytest.param(Pattern, r'^whatev.r\d$', 'whatever1', ' whatever1', id='Pattern'),
-        pytest.param(Pattern[str], r'^whatev.r\d$', 'whatever1', ' whatever1', id='Pattern[str]'),
-        pytest.param(Pattern[bytes], rb'^whatev.r\d$', b'whatever1', b' whatever1', id='Pattern[bytes]'),
-    ],
-)
-def test_pattern(pattern_type, pattern_value, matching_value, non_matching_value):
-    class Foobar(BaseModel):
-        pattern: pattern_type
-
-    f = Foobar(pattern=pattern_value)
-    assert f.pattern.__class__.__name__ == 'Pattern'
-    # check it's really a proper pattern
-    assert f.pattern.match(matching_value)
-    assert not f.pattern.match(non_matching_value)
-
-    # Check that pre-compiled patterns are accepted unchanged
-    p = re.compile(pattern_value)
-    f2 = Foobar(pattern=p)
-    assert f2.pattern is p
-
-    assert Foobar.model_json_schema() == {
-        'type': 'object',
-        'title': 'Foobar',
-        'properties': {'pattern': {'type': 'string', 'format': 'regex', 'title': 'Pattern'}},
-        'required': ['pattern'],
-    }
-
-
-@pytest.mark.parametrize(
-    'use_field',
-    [pytest.param(True, id='Field'), pytest.param(False, id='constr')],
-)
-def test_compiled_pattern_in_field(use_field):
-    """
-    https://github.com/pydantic/pydantic/issues/9052
-    https://github.com/pydantic/pydantic/pull/9053
-    """
-    pattern_value = r'^whatev.r\d$'
-    field_pattern = re.compile(pattern_value)
-
-    if use_field:
-
-        class Foobar(BaseModel):
-            str_regex: str = Field(pattern=field_pattern)
-    else:
-
-        class Foobar(BaseModel):
-            str_regex: constr(pattern=field_pattern) = ...
-
-    field_general_metadata = Foobar.model_fields['str_regex'].metadata
-    assert len(field_general_metadata) == 1
-    field_metadata_pattern = field_general_metadata[0].pattern
-
-    assert field_metadata_pattern == field_pattern
-    assert isinstance(field_metadata_pattern, re.Pattern)
-
-    matching_value = 'whatever1'
-    f = Foobar(str_regex=matching_value)
-    assert f.str_regex == matching_value
-
-    with pytest.raises(
-        ValidationError,
-        match=re.escape("String should match pattern '" + pattern_value + "'"),
-    ):
-        Foobar(str_regex=' whatever1')
-
-    assert Foobar.model_json_schema() == {
-        'type': 'object',
-        'title': 'Foobar',
-        'properties': {'str_regex': {'pattern': pattern_value, 'title': 'Str Regex', 'type': 'string'}},
-        'required': ['str_regex'],
-    }
-
-
-def test_pattern_with_invalid_param():
-    with pytest.raises(
-        PydanticSchemaGenerationError,
-        match=re.escape('Unable to generate pydantic-core schema for re.Pattern[int].'),
-    ):
-
-        class Foo(BaseModel):
-            pattern: Pattern[int]
-
-
-@pytest.mark.parametrize(
-    ('pattern_type', 'pattern_value', 'error_type', 'error_msg'),
-    [
-        pytest.param(
-            re.Pattern,
-            '[xx',
-            'pattern_regex',
-            'Input should be a valid regular expression',
-            id='re.Pattern-pattern_regex',
-        ),
-        pytest.param(re.Pattern, (), 'pattern_type', 'Input should be a valid pattern', id='re.Pattern-pattern_type'),
-        pytest.param(
-            re.Pattern[str],
-            re.compile(b''),
-            'pattern_str_type',
-            'Input should be a string pattern',
-            id='re.Pattern[str]-pattern_str_type-non_str',
-        ),
-        pytest.param(
-            re.Pattern[str],
-            b'',
-            'pattern_str_type',
-            'Input should be a string pattern',
-            id='re.Pattern[str]-pattern_str_type-bytes',
-        ),
-        pytest.param(
-            re.Pattern[str], (), 'pattern_type', 'Input should be a valid pattern', id='re.Pattern[str]-pattern_type'
-        ),
-        pytest.param(
-            re.Pattern[bytes],
-            re.compile(''),
-            'pattern_bytes_type',
-            'Input should be a bytes pattern',
-            id='re.Pattern[bytes]-pattern_bytes_type-non_bytes',
-        ),
-        pytest.param(
-            re.Pattern[bytes],
-            '',
-            'pattern_bytes_type',
-            'Input should be a bytes pattern',
-            id='re.Pattern[bytes]-pattern_bytes_type-str',
-        ),
-        pytest.param(
-            re.Pattern[bytes],
-            (),
-            'pattern_type',
-            'Input should be a valid pattern',
-            id='re.Pattern[bytes]-pattern_type',
-        ),
-    ],
-)
-def test_pattern_error(pattern_type, pattern_value, error_type, error_msg):
-    class Foobar(BaseModel):
-        pattern: pattern_type
-
-    with pytest.raises(ValidationError) as exc_info:
-        Foobar(pattern=pattern_value)
-    # insert_assert(exc_info.value.errors(include_url=False))
-    assert exc_info.value.errors(include_url=False) == [
-        {'type': error_type, 'loc': ('pattern',), 'msg': error_msg, 'input': pattern_value}
-    ]
-
-
 @pytest.mark.parametrize('validate_json', [True, False])
 def test_secretstr(validate_json):
     class Foobar(BaseModel):
@@ -1444,45 +1293,6 @@ def test_generic_without_params_error():
     ]
 
 
-def test_literal_single():
-    class Model(BaseModel):
-        a: Literal['a']
-
-    Model(a='a')
-    with pytest.raises(ValidationError) as exc_info:
-        Model(a='b')
-    # insert_assert(exc_info.value.errors(include_url=False))
-    assert exc_info.value.errors(include_url=False) == [
-        {
-            'type': 'literal_error',
-            'loc': ('a',),
-            'msg': "Input should be 'a'",
-            'input': 'b',
-            'ctx': {'expected': "'a'"},
-        }
-    ]
-
-
-def test_literal_multiple():
-    class Model(BaseModel):
-        a_or_b: Literal['a', 'b']
-
-    Model(a_or_b='a')
-    Model(a_or_b='b')
-    with pytest.raises(ValidationError) as exc_info:
-        Model(a_or_b='c')
-    # insert_assert(exc_info.value.errors(include_url=False))
-    assert exc_info.value.errors(include_url=False) == [
-        {
-            'type': 'literal_error',
-            'loc': ('a_or_b',),
-            'msg': "Input should be 'a' or 'b'",
-            'input': 'c',
-            'ctx': {'expected': "'a' or 'b'"},
-        }
-    ]
-
-
 @pytest.mark.parametrize(
     'input_value,output,human_bin,human_dec,human_sep',
     (
@@ -1563,163 +1373,6 @@ def test_bytesize_raises():
 
     with pytest.raises(PydanticCustomError, match='byte unit'):
         m.size.to('1ZiB')
-
-
-def test_default_union_types():
-    class DefaultModel(BaseModel):
-        v: int | bool | str
-
-    # do it this way since `1 == True`
-    assert repr(DefaultModel(v=True).v) == 'True'
-    assert repr(DefaultModel(v=1).v) == '1'
-    assert repr(DefaultModel(v='1').v) == "'1'"
-
-    assert DefaultModel.model_json_schema() == {
-        'title': 'DefaultModel',
-        'type': 'object',
-        'properties': {'v': {'title': 'V', 'anyOf': [{'type': t} for t in ('integer', 'boolean', 'string')]}},
-        'required': ['v'],
-    }
-
-
-def test_default_union_types_left_to_right():
-    class DefaultModel(BaseModel):
-        v: Annotated[int | bool | str, Field(union_mode='left_to_right')]
-
-    print(DefaultModel.__pydantic_core_schema__)
-
-    # int will coerce everything in left-to-right mode
-    assert repr(DefaultModel(v=True).v) == '1'
-    assert repr(DefaultModel(v=1).v) == '1'
-    assert repr(DefaultModel(v='1').v) == '1'
-
-    assert DefaultModel.model_json_schema() == {
-        'title': 'DefaultModel',
-        'type': 'object',
-        'properties': {'v': {'title': 'V', 'anyOf': [{'type': t} for t in ('integer', 'boolean', 'string')]}},
-        'required': ['v'],
-    }
-
-
-def test_union_enum_int_left_to_right():
-    class BinaryEnum(IntEnum):
-        ZERO = 0
-        ONE = 1
-
-    # int will win over enum in this case
-    assert TypeAdapter(BinaryEnum | int).validate_python(0) is not BinaryEnum.ZERO
-
-    # in left to right mode, enum will validate successfully and take precedence
-    assert (
-        TypeAdapter(Annotated[BinaryEnum | int, Field(union_mode='left_to_right')]).validate_python(0)
-        is BinaryEnum.ZERO
-    )
-
-
-def test_union_uuid_str_left_to_right():
-    IdOrSlug = UUID | str
-
-    # in smart mode JSON and python are currently validated differently in this
-    # case, because in Python this is a str but in JSON a str is also a UUID
-    assert TypeAdapter(IdOrSlug).validate_json('"f4fe10b4-e0c8-4232-ba26-4acd491c2414"') == UUID(
-        'f4fe10b4-e0c8-4232-ba26-4acd491c2414'
-    )
-    assert (
-        TypeAdapter(IdOrSlug).validate_python('f4fe10b4-e0c8-4232-ba26-4acd491c2414')
-        == 'f4fe10b4-e0c8-4232-ba26-4acd491c2414'
-    )
-
-    IdOrSlugLTR = Annotated[UUID | str, Field(union_mode='left_to_right')]
-
-    # in left to right mode both JSON and python are validated as UUID
-    assert TypeAdapter(IdOrSlugLTR).validate_json('"f4fe10b4-e0c8-4232-ba26-4acd491c2414"') == UUID(
-        'f4fe10b4-e0c8-4232-ba26-4acd491c2414'
-    )
-    assert TypeAdapter(IdOrSlugLTR).validate_python('f4fe10b4-e0c8-4232-ba26-4acd491c2414') == UUID(
-        'f4fe10b4-e0c8-4232-ba26-4acd491c2414'
-    )
-
-
-def test_default_union_class():
-    class A(BaseModel):
-        x: str
-
-    class B(BaseModel):
-        x: str
-
-    class Model(BaseModel):
-        y: A | B
-
-    assert isinstance(Model(y=A(x='a')).y, A)
-    assert isinstance(Model(y=B(x='b')).y, B)
-
-
-@pytest.mark.parametrize('max_length', [10, None])
-def test_union_subclass(max_length: int | None):
-    class MyStr(str): ...
-
-    class Model(BaseModel):
-        x: int | Annotated[str, Field(max_length=max_length)]
-
-    v = Model(x=MyStr('1')).x
-    assert type(v) is str
-    assert v == '1'
-
-
-def test_union_compound_types():
-    class Model(BaseModel):
-        values: dict[str, str] | list[str] | dict[str, list[str]]
-
-    assert Model(values={'L': '1'}).model_dump() == {'values': {'L': '1'}}
-    assert Model(values=['L1']).model_dump() == {'values': ['L1']}
-    assert Model(values=('L1',)).model_dump() == {'values': ['L1']}
-    assert Model(values={'x': ['pika']}) != {'values': {'x': ['pika']}}
-    assert Model(values={'x': ('pika',)}).model_dump() == {'values': {'x': ['pika']}}
-    with pytest.raises(ValidationError) as e:
-        Model(values={'x': {'a': 'b'}})
-    # insert_assert(e.value.errors(include_url=False))
-    assert e.value.errors(include_url=False) == [
-        {
-            'type': 'string_type',
-            'loc': ('values', 'dict[str,str]', 'x'),
-            'msg': 'Input should be a valid string',
-            'input': {'a': 'b'},
-        },
-        {
-            'type': 'list_type',
-            'loc': ('values', 'list[str]'),
-            'msg': 'Input should be a valid list',
-            'input': {'x': {'a': 'b'}},
-        },
-        {
-            'type': 'list_type',
-            'loc': ('values', 'dict[str,list[str]]', 'x'),
-            'msg': 'Input should be a valid list',
-            'input': {'a': 'b'},
-        },
-    ]
-
-
-def test_smart_union_compounded_types_edge_case():
-    class Model(BaseModel):
-        x: list[str] | list[int]
-
-    assert Model(x=[1, 2]).x == [1, 2]
-    assert Model(x=['1', '2']).x == ['1', '2']
-    assert Model(x=[1, '2']).x == [1, 2]
-
-
-def test_union_typeddict():
-    class Dict1(TypedDict):
-        foo: str
-
-    class Dict2(TypedDict):
-        bar: str
-
-    class M(BaseModel):
-        d: Dict2 | Dict1
-
-    assert M(d=dict(foo='baz')).d == {'foo': 'baz'}
 
 
 def test_custom_generic_containers():
@@ -2191,21 +1844,6 @@ def test_transform_schema_for_third_party_class():
     ]
 
 
-@pytest.mark.parametrize(
-    'typing_literal',
-    [
-        pytest.param(typing.Literal, id='typing.Literal'),
-        pytest.param(typing_extensions.Literal, id='typing_extensions.Literal'),
-    ],
-)
-def test_literal_field(typing_literal):
-
-    class Model(BaseModel):
-        foo: typing_literal['foo']  # noqa: F821
-
-    assert Model(foo='foo').foo == 'foo'
-
-
 def test_instance_of_annotation():
     class Model(BaseModel):
         # Note: the generic parameter gets ignored by runtime validation
@@ -2537,55 +2175,6 @@ def test_get_pydantic_core_schema_marker_unrelated_type() -> None:
     assert ta.validate_python('1') == 3
 
 
-def test_union_tags_in_errors():
-    DoubledList = Annotated[list[int], AfterValidator(lambda x: x * 2)]
-    StringsMap = dict[str, str]
-
-    adapter = TypeAdapter(DoubledList | StringsMap)
-
-    with pytest.raises(ValidationError) as exc_info:
-        adapter.validate_python(['a'])
-
-    # yuck
-    assert '2 validation errors for union[function-after[<lambda>(), list[int]],dict[str,str]]' in str(exc_info)
-    # the loc's are bad here:
-    assert exc_info.value.errors(include_url=False) == [
-        {
-            'input': 'a',
-            'loc': ('function-after[<lambda>(), list[int]]', 0),
-            'msg': 'Input should be a valid integer, unable to parse string as an integer',
-            'type': 'int_parsing',
-        },
-        {
-            'input': ['a'],
-            'loc': ('dict[str,str]',),
-            'msg': 'Input should be a valid dictionary',
-            'type': 'dict_type',
-        },
-    ]
-
-    tag_adapter = TypeAdapter(Annotated[DoubledList, Tag('DoubledList')] | Annotated[StringsMap, Tag('StringsMap')])
-    with pytest.raises(ValidationError) as exc_info:
-        tag_adapter.validate_python(['a'])
-
-    assert '2 validation errors for union[DoubledList,StringsMap]' in str(exc_info)  # nice
-    # the loc's are good here:
-    assert exc_info.value.errors(include_url=False) == [
-        {
-            'input': 'a',
-            'loc': ('DoubledList', 0),
-            'msg': 'Input should be a valid integer, unable to parse string as an integer',
-            'type': 'int_parsing',
-        },
-        {
-            'input': ['a'],
-            'loc': ('StringsMap',),
-            'msg': 'Input should be a valid dictionary',
-            'type': 'dict_type',
-        },
-    ]
-
-
 def test_json_value():
     adapter = TypeAdapter(JsonValue)
     valid_json_data = {'a': {'b': {'c': 1, 'd': [2, None]}}}
@@ -2670,102 +2259,6 @@ def test_on_error_omit_top_level() -> None:
     # if it hits the top level, but this documents the current behavior at least
     with pytest.raises(SchemaError, match='Uncaught Omit error'):
         ta.validate_python('a')
-
-
-@pytest.mark.skipif(
-    platform.python_implementation() == 'PyPy',
-    reason='PyPy has a bug in complex string parsing. A fix is implemented but not yet released.',
-)
-def test_complex_field():
-    class Model(BaseModel):
-        number: complex
-
-    m = Model(number=complex(1, 2))
-    assert repr(m) == 'Model(number=(1+2j))'
-    assert m.model_dump() == {'number': complex(1, 2)}
-    assert m.model_dump_json() == '{"number":"1+2j"}'
-
-    # Complex numbers presented as strings are also acceptable
-    m = Model(number='1+2j')
-    assert repr(m) == 'Model(number=(1+2j))'
-    assert m.model_dump() == {'number': complex(1, 2)}
-    assert m.model_dump_json() == '{"number":"1+2j"}'
-
-    # The part that is 0 can be omitted
-    m = Model(number='1')
-    assert repr(m) == 'Model(number=(1+0j))'
-    assert m.model_dump() == {'number': complex(1, 0)}
-    assert m.model_dump_json() == '{"number":"1+0j"}'
-
-    m = Model(number='1j')
-    assert repr(m) == 'Model(number=1j)'
-    assert m.model_dump() == {'number': complex(0, 1)}
-    assert m.model_dump_json() == '{"number":"1j"}'
-
-    m = Model(number='0')
-    assert repr(m) == 'Model(number=0j)'
-    assert m.model_dump() == {'number': complex(0, 0)}
-    assert m.model_dump_json() == '{"number":"0j"}'
-
-    m = Model(number='infj')
-    assert repr(m) == 'Model(number=infj)'
-    assert m.model_dump() == {'number': complex(0, float('inf'))}
-    assert m.model_dump_json() == '{"number":"infj"}'
-
-    m = Model(number='-nanj')
-    assert repr(m) == 'Model(number=nanj)'
-    d = m.model_dump()
-    assert d['number'].real == 0
-    assert math.isnan(d['number'].imag)
-    assert m.model_dump_json() == '{"number":"NaNj"}'
-
-    # strings with brackets and space characters are allowed as long as
-    # they follow the rule
-    m = Model(number='\t( -1.23+4.5J )\n')
-    assert repr(m) == 'Model(number=(-1.23+4.5j))'
-    assert m.model_dump() == {'number': complex(-1.23, 4.5)}
-    assert m.model_dump_json() == '{"number":"-1.23+4.5j"}'
-
-    # int and float are also accepted (with imaginary part == 0)
-    m = Model(number=2)
-    assert repr(m) == 'Model(number=(2+0j))'
-    assert m.model_dump() == {'number': complex(2, 0)}
-    assert m.model_dump_json() == '{"number":"2+0j"}'
-
-    m = Model(number=1.5)
-    assert repr(m) == 'Model(number=(1.5+0j))'
-    assert m.model_dump() == {'number': complex(1.5, 0)}
-    assert m.model_dump_json() == '{"number":"1.5+0j"}'
-
-    # Empty strings are not allowed
-    with pytest.raises(ValidationError):
-        Model(number='')
-    with pytest.raises(ValidationError):
-        Model(number='foo')
-    # Bracket missing
-    with pytest.raises(ValidationError):
-        Model(number='\t( -1.23+4.5J \n')
-    # Space between numbers
-    with pytest.raises(ValidationError):
-        Model(number='\t( -1.23 +4.5J \n')
-
-
-def test_strict_complex_field():
-    class Model(BaseModel):
-        # Only complex objects are accepted
-        number: Annotated[complex, Field(strict=True)]
-
-    m = Model(number=complex(1, 2))
-    assert repr(m) == 'Model(number=(1+2j))'
-    assert m.model_dump() == {'number': complex(1, 2)}
-    assert m.model_dump_json() == '{"number":"1+2j"}'
-
-    with pytest.raises(ValidationError):
-        m = Model(number='1+2j')
-    with pytest.raises(ValidationError):
-        m = Model(number=1.0)
-    with pytest.raises(ValidationError):
-        m = Model(number=5)
 
 
 @pytest.mark.skipif(not email_validator, reason='email_validator not installed')
@@ -2930,47 +2423,5 @@ def test_custom_serializer_override_secret_str() -> None:
 
     u = User(name='sam', password='hi')
     assert u.model_dump()['password'] == 'secret: **********'
-
-
-def test_union_respects_local_strict() -> None:
-    class MyBaseModel(BaseModel):
-        model_config = ConfigDict(strict=True)
-
-    class Model(MyBaseModel):
-        a: int | Annotated[tuple[int, int], Strict(False)] = Field(default=(1, 2))
-
-    m = Model(a=[1, 2])
-    assert m.a == (1, 2)
-
-
-def test_union_abc() -> None:
-    """https://github.com/pydantic/pydantic/issues/12230"""
-
-    class ABC_1(BaseModel, ABC):
-        @abstractmethod
-        def do_something(self):
-            pass
-
-    class ABC_2(BaseModel, ABC):
-        @abstractmethod
-        def do_something_else(self):
-            pass
-
-    class A1(ABC_1):
-        def do_something(self):
-            print('Doing something in A1')
-
-    class A2(ABC_2):
-        def do_something_else(self):
-            print('Doing something else in A2')
-
-    class X(BaseModel):
-        x: ABC_1 | ABC_2
-
-    a1 = A1()
-    a2 = A2()
-
-    X(x=a1)
-    X(x=a2)
